requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.10.6
python-multipart==0.0.9

# Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
cryptography==42.0.5
python-dotenv==1.0.1

# ML and Analytics
numpy==2.2.3
scikit-learn==1.6.1

# Monitoring and System
psutil==5.9.8
prometheus-client==0.20.0
structlog==24.1.0

# Configuration and Schema
pyyaml==6.0.1
schema==0.7.5

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
httpx==0.27.0